+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - `preferSoftwareDecodingForAds` in `TweaksConfig` to use software decoding for ads, which can be useful for low-end Android devices
+- `hideFirstFrame` in `PlayerViewConfig` to ensure no frame of the main content is shown before pre-roll ads
 
 ### Changed